Promote exponential histogram data interfaces to stable package
