+++ docs/apidiffs/current_vs_latest/opentelemetry-sdk-metrics.txt
@@ -1,4 +1,35 @@
 Comparing source compatibility of  against 
++++  NEW INTERFACE: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.metrics.data.ExponentialHistogramBuckets  (not serializable)
+	+++  CLASS FILE FORMAT VERSION: 52.0 <- n.a.
+	+++  NEW SUPERCLASS: java.lang.Object
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) java.util.List getBucketCounts()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) int getOffset()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) int getScale()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) long getTotalCount()
++++  NEW INTERFACE: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.metrics.data.ExponentialHistogramData  (not serializable)
+	+++  CLASS FILE FORMAT VERSION: 52.0 <- n.a.
+	+++  NEW INTERFACE: io.opentelemetry.sdk.metrics.data.Data
+	+++  NEW SUPERCLASS: java.lang.Object
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.metrics.data.AggregationTemporality getAggregationTemporality()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) java.util.Collection getPoints()
++++  NEW INTERFACE: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.metrics.data.ExponentialHistogramPointData  (not serializable)
+	+++  CLASS FILE FORMAT VERSION: 52.0 <- n.a.
+	+++  NEW INTERFACE: io.opentelemetry.sdk.metrics.data.PointData
+	+++  NEW SUPERCLASS: java.lang.Object
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) long getCount()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) java.util.List getExemplars()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) double getMax()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) double getMin()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.metrics.data.ExponentialHistogramBuckets getNegativeBuckets()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) io.opentelemetry.sdk.metrics.data.ExponentialHistogramBuckets getPositiveBuckets()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) int getScale()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) double getSum()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) long getZeroCount()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) boolean hasMax()
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) boolean hasMin()
+***  MODIFIED INTERFACE: PUBLIC ABSTRACT io.opentelemetry.sdk.metrics.data.MetricData  (not serializable)
+	===  CLASS FILE FORMAT VERSION: 52.0 <- 52.0
+	+++  NEW METHOD: PUBLIC(+) io.opentelemetry.sdk.metrics.data.ExponentialHistogramData getExponentialHistogramData()
 ***  MODIFIED INTERFACE: PUBLIC ABSTRACT io.opentelemetry.sdk.metrics.export.MetricReader  (not serializable)
 	===  CLASS FILE FORMAT VERSION: 52.0 <- 52.0
 	+++  NEW INTERFACE: java.io.Closeable
