+++ sdk/metrics-testing/src/test/java/io/opentelemetry/sdk/testing/assertj/MetricAssertionsTest.java
@@ -13,9 +13,9 @@ import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
 import io.opentelemetry.sdk.metrics.data.AggregationTemporality;
 import io.opentelemetry.sdk.metrics.data.MetricData;
+import io.opentelemetry.sdk.metrics.internal.data.ImmutableExponentialHistogramData;
 import io.opentelemetry.sdk.metrics.internal.data.ImmutableGaugeData;
 import io.opentelemetry.sdk.metrics.internal.data.ImmutableMetricData;
-import io.opentelemetry.sdk.metrics.internal.data.exponentialhistogram.ExponentialHistogramData;
 import io.opentelemetry.sdk.resources.Resource;
 import java.util.Collections;
 import org.junit.jupiter.api.Test;
@@ -33,7 +33,7 @@ public class MetricAssertionsTest {
           /* name= */ "exponential_histogram",
           /* description= */ "description",
           /* unit= */ "unit",
-          ExponentialHistogramData.create(
+          ImmutableExponentialHistogramData.create(
               AggregationTemporality.CUMULATIVE,
               // Points
               Collections.emptyList()));
@@ -45,7 +45,7 @@ public class MetricAssertionsTest {
           /* name= */ "exponential_histogram_delta",
           /* description= */ "description",
           /* unit= */ "unit",
-          ExponentialHistogramData.create(
+          ImmutableExponentialHistogramData.create(
               AggregationTemporality.DELTA,
               // Points
               Collections.emptyList()));
