+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/data/ExponentialHistogramBuckets.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.metrics.internal.data.exponentialhistogram;
+package io.opentelemetry.sdk.metrics.data;
 
 import java.util.List;
 import javax.annotation.concurrent.Immutable;
@@ -20,8 +20,7 @@ import javax.annotation.concurrent.Immutable;
  * <code>offset</code> is -3, the bucket lower bounds would be 0.125, 0.25, 0.5, 1.0, 2,0, etc. If
  * <code>offset</code> is +3, the bucket lower bounds would be 8.0, 16.0, 32.0, etc.
  *
- * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
- * at any time.
+ * @since 1.23.0
  */
 @Immutable
 public interface ExponentialHistogramBuckets {
