+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/data/ImmutableExponentialHistogramData.java
@@ -3,28 +3,28 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.metrics.internal.data.exponentialhistogram;
+package io.opentelemetry.sdk.metrics.internal.data;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.sdk.metrics.data.AggregationTemporality;
+import io.opentelemetry.sdk.metrics.data.ExponentialHistogramData;
+import io.opentelemetry.sdk.metrics.data.ExponentialHistogramPointData;
 import java.util.Collection;
 import java.util.Collections;
+import javax.annotation.concurrent.Immutable;
 
 /**
- * A simple, autovalue implementation of {@link ExponentialHistogramData}. For more detailed javadoc
- * on the type, see {@link ExponentialHistogramData}.
- *
- * <p>See:
- * https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/datamodel.md#exponentialhistogram
+ * Auto value implementation of {@link ExponentialHistogramData}.
  *
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
  */
+@Immutable
 @AutoValue
 public abstract class ImmutableExponentialHistogramData implements ExponentialHistogramData {
 
   private static final ExponentialHistogramData EMPTY =
-      ExponentialHistogramData.create(AggregationTemporality.CUMULATIVE, Collections.emptyList());
+      create(AggregationTemporality.CUMULATIVE, Collections.emptyList());
 
   public static ExponentialHistogramData empty() {
     return EMPTY;
@@ -37,10 +37,4 @@ public abstract class ImmutableExponentialHistogramData implements ExponentialHi
   }
 
   ImmutableExponentialHistogramData() {}
-
-  @Override
-  public abstract AggregationTemporality getAggregationTemporality();
-
-  @Override
-  public abstract Collection<ExponentialHistogramPointData> getPoints();
 }
