--- sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/data/exponentialhistogram/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** The data format to model exponential histograms for export. */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.metrics.internal.data.exponentialhistogram;
-
-import javax.annotation.ParametersAreNonnullByDefault;
